+++ src/test/scala/com/guizmaii/scalajwt/Generators.scala
@@ -13,6 +13,8 @@ import java.util.UUID
 
 object Generators {
 
+  val nonEmptyStringGen: Gen[String] = Gen.alphaStr.filter(_.trim.length > 0)
+
   def jwkGen(keyPair: KeyPair): Gen[JWK] = Gen.const {
     new RSAKey.Builder(keyPair.getPublic.asInstanceOf[RSAPublicKey])
       .privateKey(keyPair.getPrivate.asInstanceOf[RSAPrivateKey])