+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- **Handle the case where the JWT token is invalid**
+- **Cover ConfigurableJwtValidator with tests**
 - **Update scalafmt**
 - **Update Nimbus from v4.39.2 to v4.41**
 - **Update Scala from v2.12.2 to v2.12.3**